refactor: separate out build quickpicks
